requirements.txt
cryptography==36.0.2
alembic
beautifulsoup4
Flask
Flask_Bcrypt
Flask_Login
Flask-Dance
Flask-Dance[sqla]
flask_cache
Flask_Migrate
flask_saml2
flask_sqlalchemy
flask_wtf
langchain
pytest
SQLAlchemy
Werkzeug
WTForms
Flask_Session
psycopg2-binary
sentence-transformers
chromadb
openai
langchain-openai
pypdf
rapidocr-onnxruntime
waitress
pyhtml2pdf
python-dotenv
langchain_community